feat(network-activity-plugin): support SSE (#45)
Server-side events are now supported in the Network Activity plugin. As long as you have react-native-sse installed in your project, the plugin will detect the connection and intercept all events, displaying them in the DevTools UI. A small change has been made to the @rozenite/vite-plugin, allowing the use of the manualChunks property when building the React Native bundle.